dev-codacy

+++ tests/test_regions.py
@@ -13,12 +13,12 @@ def test_radec2sky():
     """Test function: Region.radec2sky"""
     ra, dec = (15, -45)
     sky = Region.radec2sky(ra, dec)
-    assert np.all(sky == [(ra, dec)]), "radec2sky broken on non-list input"
+    if not (np.all(sky == [(ra, dec)])): raise AssertionError("radec2sky broken on non-list input")
     ra = [0, 10]
     dec = [-45, 45]
     sky = Region.radec2sky(ra, dec)
     answer = np.array([(ra[0], dec[0]), (ra[1], dec[1])])
-    assert np.all(sky == answer), 'radec2sky broken on list input'
+    if not (np.all(sky == answer)): raise AssertionError('radec2sky broken on list input')
 
 
 def test_sky2ang_symmetric():
@@ -28,7 +28,7 @@ def test_sky2ang_symmetric():
     tp = np.array([[tp[0][1], tp[0][0]]])
     sky2 = Region.sky2ang(tp)
     sky2 = np.array([[sky2[0][1], sky2[0][0]]])
-    assert np.all(abs(sky-sky2) < 1e-9), "sky2ang failed to be symmetric"
+    if not (np.all(abs(sky-sky2) < 1e-9)): raise AssertionError("sky2ang failed to be symmetric")
 
 
 def test_sky2ang_corners():
@@ -36,7 +36,7 @@ def test_sky2ang_corners():
     corners = np.radians([[0, 0], [360, -90]])
     theta_phi = Region.sky2ang(corners)
     answers = np.array([[np.pi/2, 0], [np.pi, 2*np.pi]])
-    assert np.all(theta_phi - answers < 1e-9), 'sky2ang corner cases failed'
+    if not (np.all(theta_phi - answers < 1e-9)): raise AssertionError('sky2ang corner cases failed')
 
 
 def test_sky2vec_corners():
@@ -44,7 +44,7 @@ def test_sky2vec_corners():
     sky = np.radians([[0, 0], [90, 90], [45, -90]])
     answers = np.array([[1, 0, 0], [0, 0, 1], [0, 0, -1]])
     vec = Region.sky2vec(sky)
-    assert np.all(vec - answers < 1e-9), 'sky2vec corner cases failed'
+    if not (np.all(vec - answers < 1e-9)): raise AssertionError('sky2vec corner cases failed')
 
 
 def test_vec2sky_corners():
@@ -52,7 +52,7 @@ def test_vec2sky_corners():
     vectors = np.array([[1, 0, 0], [0, 0, 1], [0, 0, -1]])
     skycoords = Region.vec2sky(vectors, degrees=True)
     answers = np.array([[0, 0], [0, 90], [0, -90]])
-    assert np.all(skycoords == answers), 'vec2sky fails on corners'
+    if not (np.all(skycoords == answers)): raise AssertionError('vec2sky fails on corners')
 
 
 def test_sky2vec2sky():
@@ -61,9 +61,9 @@ def test_sky2vec2sky():
     sky = Region.radec2sky(ra, dec)
     vec = Region.sky2vec(sky)
     sky2 = Region.vec2sky(vec)
-    assert np.all(np.array(sky2) - np.array(sky) == 0), "sky2vec2sky failed"
+    if not (np.all(np.array(sky2) - np.array(sky) == 0)): raise AssertionError("sky2vec2sky failed")
     vec2 = Region.sky2vec(sky2)
-    assert np.all(np.array(vec) - np.array(vec2) == 0), 'vec2sky2vec failed'
+    if not (np.all(np.array(vec) - np.array(vec2) == 0)): raise AssertionError('vec2sky2vec failed')
 
 
 def test_add_circles_list_scalar():
@@ -84,7 +84,7 @@ def test_add_circles_list_scalar():
     for i in range(1, region1.maxdepth+1):
         if len(region1.pixeldict[i].difference(region2.pixeldict[i])) > 0:
             test = False
-    assert test, 'add_circles gives different results for lists and scalars'
+    if not (test): raise AssertionError('add_circles gives different results for lists and scalars')
 
 
 def test_renorm_demote_symmetric():
@@ -103,7 +103,7 @@ def test_renorm_demote_symmetric():
     for i in range(1, region.maxdepth+1):
         if len(end_dict[i].difference(start_dict[i])) > 0:
             test = False
-    assert test, 'renorm and demote are not symmetric'
+    if not (test): raise AssertionError('renorm and demote are not symmetric')
 
 
 def test_sky_within():
@@ -113,9 +113,9 @@ def test_sky_within():
     radius = np.radians([0.1, 0.1])
     region = Region(maxdepth=11)
     region.add_circles(ra, dec, radius)
-    assert np.all(region.sky_within(ra[0], dec[0])), "Failed on position at center of region"
-    assert np.all(region.sky_within(ra, dec)), "Failed on list of positions"
-    assert not np.any(region.sky_within(ra[0]+5*radius[0], dec[0])), "Failed on position outside of region"
+    if not (np.all(region.sky_within(ra[0], dec[0]))): raise AssertionError("Failed on position at center of region")
+    if not (np.all(region.sky_within(ra, dec))): raise AssertionError("Failed on list of positions")
+    if np.any(region.sky_within(ra[0]+5*radius[0], dec[0])): raise AssertionError("Failed on position outside of region")
 
 
 def test_pickle():
@@ -131,7 +131,7 @@ def test_pickle():
         import pickle
     pickle.dump(region, open('out_temp.mim', 'wb'))
     region2 = pickle.load(open('out_temp.mim','rb'))
-    assert region.pixeldict == region2.pixeldict, 'pickle/unpickle does not give same region'
+    if not (region.pixeldict == region2.pixeldict): raise AssertionError('pickle/unpickle does not give same region')
     os.remove('out_temp.mim')
 
 
@@ -146,7 +146,7 @@ def test_reg():
     region = Region(maxdepth=5)
     region.add_circles(ra, dec, radius)
     region.write_reg('test.reg')
-    assert os.path.exists('test.reg')
+    if not (os.path.exists('test.reg')): raise AssertionError()
     os.remove('test.reg')
 
 
@@ -160,7 +160,7 @@ def test_poly():
     positions = list(zip(np.radians(ra), np.radians(dec)))
     region.add_poly(positions)
     region.write_reg('test.reg')
-    assert os.path.exists('test.reg')
+    if not (os.path.exists('test.reg')): raise AssertionError()
     os.remove('test.reg')
 
 
@@ -169,7 +169,7 @@ def test_write_fits():
     a = Region()
     a.add_circles(12, 0, 0.1)
     a.write_fits('test_MOC.fits')
-    assert os.path.exists('test_MOC.fits')
+    if not (os.path.exists('test_MOC.fits')): raise AssertionError()
     os.remove('test_MOC.fits')
 
 
@@ -183,7 +183,7 @@ def test_without():
     b = Region(maxdepth=7)
     b.add_circles(0, np.radians(-90), np.radians(0.5))
     a.without(b)
-    assert a.get_area() <= (area - b.get_area()), "test_without FAILED"
+    if not (a.get_area() <= (area - b.get_area())): raise AssertionError("test_without FAILED")
 
 
 def test_intersect():
@@ -195,7 +195,7 @@ def test_intersect():
     b = Region(maxdepth=7)
     b.add_circles(0, np.radians(-90), np.radians(0.5))
     a.intersect(b)
-    assert a.get_area() == b.get_area(), "test_intersect FAILED"
+    if not (a.get_area() == b.get_area()): raise AssertionError("test_intersect FAILED")
 
 
 def test_demote():
@@ -203,9 +203,9 @@ def test_demote():
     a.add_circles(0, np.radians(-90), np.radians(1))
     ipd = a.pixeldict.copy()
     fpd = a.get_demoted()
-    assert fpd == a.pixeldict[8]
+    if not (fpd == a.pixeldict[8]): raise AssertionError()
     for i in range(1, 8):
-        assert len(a.pixeldict[i]) == 0
+        if not (len(a.pixeldict[i]) == 0): raise AssertionError()
 
 
 def test_symmetric_difference():
@@ -218,7 +218,7 @@ def test_symmetric_difference():
     b = Region(maxdepth=7)
     b.add_circles(0, np.radians(-90), np.radians(0.5))
     a.symmetric_difference(b)
-    assert a.get_area() == area - b.get_area(), "test_symmetric_difference FAILED"
+    if not (a.get_area() == area - b.get_area()): raise AssertionError("test_symmetric_difference FAILED")
 
 
 if __name__ == "__main__":
